[FEATURE] Add API to get server configuration
[DESC.]
- add API `GET:/webida/api/app/configs` for getting information of system apps and feature enables.
- seperate `App` model from `app-manager` module
- remove useless package, `connect-mongo`, from `pacakge.json`
- add a feature enabling JSONP response for this new API
- fix some bugs
