+++ src/server/conf/default-conf.js
@@ -175,6 +175,7 @@ var conf = {
     },
 
     guestMode: {
+        enable: false,
         accountPrefix: '__webida-guest-',
         passwordPrefix: 'qlalf!',
         ttl: 24 * 60 * 60 * 2 // 48 hours
