+++ src/server/package.json
@@ -9,7 +9,6 @@
         "express-session": "1.9.1",
         "http-proxy": "0.10.3",
         "http-master": "~1.0.18",
-        "connect-mongo": "0.4.2",
         "session-file-store": "0.0.12",
         "connect-sqlite3": "0.9.5",
         "socket.io": "1.2.0",
@@ -53,7 +52,6 @@
         "passport-github": "*",
         "passport-google-oauth": "*",
         "nodemailer": "0.7.x",
-        "connect-mongo": "*",
         "mysql": "2.5.3",
         "line-reader": "0.2.3",
         "filequeue": "0.5.0",
